fix: port type

+++ migrations/2023-12-21-135159_add_indexer_status_port/up.sql
@@ -1,3 +1,3 @@
 -- Your SQL goes here
 ALTER TABLE indexers
-ADD COLUMN status_server_port SMALLINT DEFAULT 1234;
+ADD COLUMN status_server_port INTEGER DEFAULT 1234;
