+++ src/infra/repositories/indexer_repository.rs
@@ -22,7 +22,7 @@ pub struct IndexerDb {
     pub process_id: Option<i64>,
     pub target_url: Option<String>,
     pub table_name: Option<String>,
-    pub status_server_port: Option<i16>,
+    pub status_server_port: Option<i32>,
 }
 
 #[derive(Deserialize)]
@@ -38,7 +38,7 @@ pub struct NewIndexerDb {
     pub type_: String,
     pub target_url: Option<String>,
     pub table_name: Option<String>,
-    pub status_server_port: Option<i16>,
+    pub status_server_port: Option<i32>,
 }
 
 #[derive(Deserialize, Insertable)]
